requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
colorzero==2.0
docutils==0.19
gpiozero==1.6.2
idna==3.4
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
Pillow==9.2.0
Pygments==2.13.0
requests==2.28.1
urllib3==1.26.12
evdev
usb
pycups
qrcode
pyclean